% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{as.numeric.factor}
\alias{as.numeric.factor}
\title{as.numeric.factor}
\usage{
\method{as.numeric}{factor}(x)
}
\arguments{
\item{x}{numeric value}
}
\description{
Convert numeric to factor
}
\examples{
as.numeric.factor(x)
}
\keyword{factor}
