% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{basic.setup}
\alias{basic.setup}
\title{basic.setup}
\usage{
basic.setup(genecounts, metadata, genemap, removePC = TRUE,
  read_filter = 10000, nonzero_filter = 0.5, equalize = FALSE)
}
\arguments{
\item{genecounts}{matrix; the matrix of gene counts (samples in columns, genes in rows)}

\item{metadata}{matrix; the metadata matrix (samples in rows, data in columns)}

\item{genemap}{matrix; the matrix of ensembl to hgnc genes}

\item{removePC}{boolean; boolean variable indicating whether or not to remove non-protein-coding genes (default = TRUE)}

\item{read_filter}{integer; threshold for number of reads required to include a sample (default = 10,000)}

\item{nonzero_filter}{float; value indicating the proportion of samples that must have non-zero gene expression value for a gene to be included (detault = .5)}

\item{equalize}{boolean; a variable indicating whether to subsample the dataset to create equal classes (NO LONGER IN USE)}
}
\description{
Take matrix of gene counts and metadata and transform them into filtered, ExpressionSet object to be used in downstream analysis.
}
\examples{
basic.setup(genecounts, metadata, removePC = TRUE, read_Filter = 10000, nonzero_filter = .5, equalize=FALSE)
}
