% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_proportions}
\alias{cell_proportions}
\title{cell_proportions}
\format{A matrix with rows corresponding to sample IDs and columns corresponding to cell types}
\usage{
cell_proportions
}
\description{
A matrix of estimated cell type proportions estimated from EpiDISH
}
\keyword{datasets}
