% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{classification_eset}
\alias{classification_eset}
\title{classification_eset}
\format{An ExpressionSet object with filtered AND NORMALIZED gene counts data in the exprs() field and metadata in the pData() field}
\usage{
classification_eset
}
\description{
This data contains the quality-filtered genecounts aligned to the mBAL metadata patient IDs for only group 1 and group 4 patients;
This includes all features for training on. SCALING has been applied to all metadata features
}
\keyword{datasets}
