% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host_fns.R
\name{classify_using_host_transcriptome.loocv}
\alias{classify_using_host_transcriptome.loocv}
\title{classify_using_host_transcriptome.loocv}
\usage{
classify_using_host_transcriptome.loocv(eset, train, test, control,
  master_genemap, full_model = FALSE, save_PDF = FALSE, plot = FALSE,
  alpha = 0.5)
}
\arguments{
\item{eset}{ExpressionSet output from traditionalclassification.setup}

\item{train}{list of training samples}

\item{test}{List of test samples}

\item{control}{list of covariats to control}

\item{master_genemap}{matrix of HGNC to ENSEMBL ID mappings}

\item{full_model}{Boolean, whether to use standard logistic regression probabilities}

\item{plot}{Boolean, whether to plot the training features in a heatmap}

\item{alpha}{elastic net alpha parameter}

\item{save_pdf}{Boolean, whether to generate a saved PDF}
}
\value{
List of outputs including 
1. test_classification - numeric vector of classification probabilities for each test sample, 
2. subset - matrix of samples x genes selected as predictive, 
3. train_classification - numeric vector of classification probabilities for each training sample,
}
\description{
Use this method to run host transcriptome classification. 
It takes in a training and test dataset consisting of gene expression values for each individual and 
returns the performance of the host expression classifier on the test samples. 
The basis of this classificaiton function is regularized linear regression, although use of a random forest can be specified.
}
\examples{
classify_using_host_transcriptome.loocv(eset, train, test, master_genemap, full_model=FALSE, save_PDF=FALSE, plot=FALSE)
}
\keyword{classify}
\keyword{host}
