% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host_fns.R
\name{classify_using_tsalik_et_al.loocv}
\alias{classify_using_tsalik_et_al.loocv}
\title{classify_using_tsalik_et_al.loocv}
\usage{
classify_using_tsalik_et_al.loocv(genes_of_interest, eset, train, test,
  full_model = FALSE, save_PDF = FALSE, plot = FALSE)
}
\arguments{
\item{genes_of_interest}{list of genes from the original paper}

\item{eset}{ExpressionSet output from traditionalclassification.setup}

\item{train}{list of training samples}

\item{test}{List of test samples}

\item{full_model}{Boolean, whether to use standard logistic regression probabilities}

\item{plot}{Boolean, whether to plot the training features in a heatmap}

\item{save_pdf}{Boolean, whether to generate a saved PDF}
}
\value{
List of outputs including 
1. test_classification - numeric vector of classification probabilities for each test sample, 
2. subset - matrix of samples x genes selected as predictive, 
3. train_classification - numeric vector of classification probabilities for each training sample,
}
\description{
Take pre-specified gense of interest, along with a training and test set consisting of host transcriptome gene counts and 
return the performance of the host expression classifier trained on that subset of genes. This function was developed to evaluate the performance of existing 
host gene expression classifiers from the literature.
}
\examples{
classify_using_tsalik_et_al.loocv(genes_of_interest, eset, train, test, full_model = FALSE, save_PDF=FALSE, plot=FALSE)
}
\keyword{classify}
\keyword{host}
\keyword{tsalik}
