% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_fns.R
\name{cohort_summary}
\alias{cohort_summary}
\title{cohort summary}
\usage{
cohort_summary(eset)
}
\arguments{
\item{eset}{ExpressionSet object containing the genecounts and metadata for the cohort}
}
\description{
Plot a variety of summary metrics about the cohort including...
effective group ID (Grouped by clinical diagnosis into Groups 1 - 4), 
pathogen type, 
gram stain,
gender,
age
Note: the metadata plots are hard-coded to work with the metadata of this particular study.
}
\examples{
cohort_summary(eset)
}
\keyword{cohort}
