% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{filterZeroRows}
\alias{filterZeroRows}
\title{filterZeroRows}
\usage{
filterZeroRows(input_matrix, is_nonzero_in_threshold)
}
\arguments{
\item{input_matrix}{matrix; full genes x samples matrix}

\item{is_nonzero_in_threshold}{float; the proportion of samples that must contain any given gene}
}
\description{
Remove rows (genes) that are present in less than X% of samples
}
\examples{
filterZeroRows(input_matrix, is_nonzero_in_threshold)
}
\keyword{filter}
