% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{format_host_features}
\alias{format_host_features}
\title{format_host_features}
\usage{
format_host_features(eset, zscore = FALSE)
}
\arguments{
\item{eset}{ExpressionSet; the ExpressionSet object (output from basic.setup, binding metadata with genecoutns)}

\item{zscore}{boolean; boolean variable indicating whether or not to apply gene-wise z-score}
}
\description{
Take ExpressionSet and apply z-score scaling and centering of each gene
}
\examples{
format_host_features(eset, zscore = FALSE)
}
