% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fulldata_eset}
\alias{fulldata_eset}
\title{hostCD}
\format{An ExpressionSet object with filtered AND NORMALIZED gene counts data in the exprs() field and metadata in the pData() field}
\usage{
fulldata_eset
}
\description{
This data contains the quality-filtered genecounts aligned to the mBAL metadata patient IDs;
The mBAL metadata contains diversity, microbe, and microbe_vector features at this point. NO scaling has been done to metadata features.
}
\keyword{datasets}
