% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genemap}
\alias{genemap}
\title{genemap}
\format{A matrix containing ensembl gene ids and hgnc gene ids}
\usage{
genemap
}
\description{
This matrix contains a map between ensembl and HGNC gene identifiers
}
\keyword{datasets}
