% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_utility_fns.R
\name{leave_one_out.split}
\alias{leave_one_out.split}
\title{Leave One Out Split}
\usage{
leave_one_out.split(df)
}
\arguments{
\item{df}{full data matrix}
}
\description{
This function splits input data for leave-one-out cross-validation
}
\examples{
tenfold.split(myData)
}
\keyword{split}
