% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_genecounts}
\alias{raw_genecounts}
\title{Raw gene counts}
\format{A data frame with [number of genes] rows and [number of patients] columns:
\describe{
  \item{genes}{All genes}
  \item{patients}{All patients}
  ...
}}
\usage{
raw_genecounts
}
\description{
This data contains RNA-seq gene counts for all patients sequencing in mBAL study.
}
\keyword{datasets}
