% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_metadata}
\alias{raw_metadata}
\title{Raw metadata}
\format{A data frame with [number of patients] rows and [number of clinical features] columns:
\describe{
  \item{clinical features}{clinical and NGS-based metadata features collected for the study}
  \item{patients}{All patients}
  ...
}}
\usage{
raw_metadata
}
\description{
This data contains clinial metadata for all patients sequencing in mBAL study.
}
\keyword{datasets}
