% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{suarez_ensembl}
\alias{suarez_ensembl}
\title{tsalik_ensembl}
\format{A list of ensembl identifiers for genes used in the Suarez et al paper}
\usage{
suarez_ensembl
}
\description{
A list of ensembl identifiers for genes used in the Suarez et al paper
}
\keyword{datasets}
