% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{traditionalclassification.setup}
\alias{traditionalclassification.setup}
\title{traditionalclassification.setup}
\usage{
traditionalclassification.setup(eset, normalization_method = "DESeq",
  rld = FALSE)
}
\arguments{
\item{eset}{ExpressionSet; the ExpressionSet object (output from basic.setup, binding metadata with genecoutns)}

\item{normalization_method}{string; which normalization method to use, must be one of "DESeq" or "voom"; default = "DESeq"}

\item{rld}{boolean; boolean variable indicating whether or not to apply rld transformation; if not, then just return log(normalized counts)}
}
\description{
Take ExpressionSet and apply standard classification filtering
}
\examples{
traditionalclassification.setup(eset, normalization_method = "DESeq", rld = FALSE)
}
