% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tsalik_ensembl}
\alias{tsalik_ensembl}
\title{tsalik_ensembl}
\format{A list of ensembl identifiers for genes used in the Tsalik et al paper}
\usage{
tsalik_ensembl
}
\description{
A list of ensembl identifiers for genes used in the Tsalik et al paper
}
\keyword{datasets}
