% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unknown_eset}
\alias{unknown_eset}
\title{unknown_eset}
\format{An ExpressionSet object with filtered AND NORMALIZED gene counts data for group 2 and group 3 in the exprs() field and metadata in the pData() field}
\usage{
unknown_eset
}
\description{
This data contains the quality-filtered genecounts aligned to the mBAL metadata patient IDs for only group 2 and group 3 patients;
This includes all features for training on. SCALING has been applied to all metadata features
}
\keyword{datasets}
